[otbn,rtl,dv] Run secure wipe after reset
This commit lets OTBN run a secure wipe of its internal registers after
reset.  This is a prerequisite for removing the reset signal from (some
of) the internal registers, which is an upcoming change, for two
reasons:

1.  Registers that are neither reset nor wiped before code execution
    could contain state from a previous execution that got aborted by
    a reset.  If this state survives a reset, it could be leaked by a
    malicious program.

2.  Without a reset, the integrity bits of integrity-protected registers
    generally don't match the data bits, which causes any read on such a
    register to trigger an integrity error.  This is unexpected behavior
    for instructions that read, modify, and write a register.  For
    example, instructions that are guaranteed to set a register to zero,
    such as `bn.xor wN, wN, wN`, cannot be used with uninitialized
    registers because they would trigger an integrity error.  This is
    problematic because these instructions are commonly used to
    initialize registers.

The initial secure wipe is exposed to users of OTBN through the `STATUS`
register:  A new state, `INIT_SEC_WIPE` (`0x10`), has been added to that
register.  This state signals that OTBN is performing its initial secure
wipe.  As with other non-`IDLE` states, OTBN does not execute commands
while it is in the `INIT_SEC_WIPE` state.  Since the initial secure wipe
starts directly after reset, the reset value of the `STATUS` register is
now `INIT_SEC_WIPE`.  Once the secure wipe is complete, OTBN
automatically changes into the `IDLE` state.  Thus, any software that
only assumes OTBN to execute a command when it is idle (as already
written in the specification) remains compatible.

This commit also extends the DV model so that:
- It also goes through a secure wipe after reset (to match the trace
  entries the RTL tracer produces); and
- It asserts that a secure wipe has completed whenever any code is
  executed.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>