[otbn,dv] Monitor model status change going into reset
A reset can also trigger a status change.  This is not observable if the
reset happens in the same status that is the post-reset status, which is
mostly the case.  When the reset happens during a different status than the
post-reset status, however, the status of the model must be updated through
this monitor.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>