[otbn,dv] Override run_csr_vseq to wait for Idle status
OTBN is specified to ignore writes to some of its CSRs while it is not in
the Idle status.  If the random CSR read/write test writes those CSRs
while OTBN is not Idle, the test will fail.

This commit waits for OTBN to be Idle before the random read/write test of
any CSR.  This is correct but somewhat overly restrictive since some CSRs
can be written while OTBN is not Idle.  Modeling this more precisely will
become a separate issue.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>